edm_backend/requirements.txt
Flask==2.1.0
python-dotenv==0.21.1
python-docx==0.8.11
Flask-CORS==4.0.0
Werkzeug==2.2.2
selenium==4.11.2
urllib3==1.26.7